core: Enable Live Merge in Snapshots Overview
Enable Live Merge in the snapshots overview.  This required enabling
the action in CanDoAction using code similar to RemoveSnapshotCommand,
and branching RemoveDiskSnapshotsCommand to use CoCo in executing the
Live Merge rather than task handlers for cold merge.

The task handlers for cold merge are initialized early in the command
lifecycle, before the command knows if the VM is up.  Therefore, a
means of disabling these handlers is provided so that once the command
can determine a Live Merge is occurring, it can disable the handlers
and rely on command coordination for the merge execution.

Change-Id: I30d152a70ebdeacf0410d7ad15a49e44fb564a18
Bug-Url: https://bugzilla.redhat.com/1133890
Signed-off-by: Greg Padgett <gpadgett@redhat.com>