core: remove unneeded code which might produce NPE
VdsEventListener#processOnVmPoweringUp contained block of code in a if
clause that could never be executed as all the commands that implement
the IVdsAsyncCommand interface returned null in their
getAutoStartVdsId method, thus the condition in the if statement was
always false.

Because of a race which is similar to the one which was fixed by
Icaca2a3, the check inside the if-condition could produce NPE.

So in this patch this block of code is removed. The signature of
VdsEventListener#processOnVmPoweringUp and IVdsAsyncCommand interface
were cleaned up by removing things which are not needed anymore.

Change-Id: Iffef34726c7feb87be1436c003eded2e226d7e5d
Bug-Url: https://bugzilla.redhat.com/1134701
Signed-off-by: Arik Hadas <ahadas@redhat.com>