core: do not consider powering up -> up as powered up transition
If the VM status is changed from powering-up to up, we should not do the
handling we do for powered up VMs, as we already did it when the VM was
switched to powering-up state.

Checking the display as part of the check that determines whether the VM
should be handled as VM that was powered-up, is not needed and thus it
is also removed as part of this change.

Change-Id: Ifc72aa638ca5722fe5bb50c600374f47a42511cd
Signed-off-by: Arik Hadas <ahadas@redhat.com>