core: prevent incorrect down events on migration
We are not supposed to produce audit log events when VM went down on
a host which is different than the one it should run on. It should only
happen when the VM is migrating.

We used to eliminate such audit logs when the VM went down normally,
from now on we will eliminate them also when the VM went down because of
an error.

Change-Id: I90b684e75972ca8eb384c33f7fab8b73640ada20
Bug-Url: https://bugzilla.redhat.com/1104195
Signed-off-by: Arik Hadas <ahadas@redhat.com>