core: ImportVmCommand storage allocation checks
This patch is a part of a series of patches, adding storage allocation
validations to the system when they're missing, and replacing old
verification usage with unified, new, correct and tested verification.
This patch did this for ImportVmCommand, while added new space allocation
validation to storage domain validators regarding snapshoted disks to be
cloned with the spnapshots. Tests were amended accordingly.

Change-Id: Ifbb1d985f9afa476452d1d2b78be1fd18c128c8f
Bug-Url: https://bugzilla.redhat.com/1053746
Signed-off-by: Vered Volansky <vvolansk@redhat.com>