engine: Change reorder vNIC action's permissions
Changed the aforementioned action to belong to ActionGroup.CREATE_VM
rather than ActionGroup.CONFIGURE_VM_NETWORK, as it's an integral part
of the process of creating a VM (it shouldn't be called in the context
of a pre-existing VM in order to re-arrange its NICs).

This might seem askew - this whole command can (and likely should) be
a part of the Create VM command, if that is changed to receive as a
parameter a list of NICs to create.

Change-Id: Iaef9b2e9ae9258210f48b59d89467247ff9a2468
Bug-Url: https://bugzilla.redhat.com/1134554
Signed-off-by: Lior Vernia <lvernia@redhat.com>