engine: Re-assign display/migration roles to mgmt network
While we recently started marking the management network explicitly as
display and migration by default, we did not take care of the scenario
where the roles are assigned to another network which is then removed
- in which case we're left with no explicit display/migration network,
and only fall back to the management network implicitly.

Now when a network holding the role is removed, it is re-assigned to
the management network.

Change-Id: I1fe94bdfda672ddf51b075415a02a80ff634a90a
Bug-Url: https://bugzilla.redhat.com/1135785
Signed-off-by: Lior Vernia <lvernia@redhat.com>