restapi: Avoid permissions in VmBase
A recent change (commit d2aa338) refactored the XML schema to simplify
the VM and template entities using a shared base type. Unfortunately
this change makes it impossible to generate the Java SDK, as it
generates conflicting get methods for the permissions sub collection.
Changing the Java SDK code generator to avoid this is possible but risky
for 3.5, so this patch changes the XML schema moving the "permissions"
property from the shared base type to the derived VM and template types.
This isn't ideal, but avoids the issue in the generator of the Java SDK.

Change-Id: I5aed6471a4f77a6df41ebd298f35807fbec616e1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>