restapi: Replace "storage_manager" with "spm"
Currently the SPM priority of a host is modeled in the RESTAPI as an
attribute of an element that also has text content:

  <host>
    <storage_manager priority="5">true</storage_manager>
  </host>

This means that it is impossible to send an udpate containing only the
priority and not the text content. For example, a document like the
following:

  <host>
    <storage_manager priority="5"/>
  </host>

Is not valid because the text is declared as xs:boolean, and an empty
text isn't a valid value.

To avoid this issue this patch removes the "storage_manager" element
from the RSDL metadata and introduces a new "spm" element that contains
inner elements for the priority and the SPM status. For example:

  <host>
    <spm>
      <priority>5</priority>
      <status>
        <state>spm</state>
      </status>
    </spm>
  </host>

This way clients can send updates for only the priority:

  <host>
    <spm>
      <priority>6</priority>
    </spm>
  </host>

The previous "storage_manager" element and its behaviour are preserved
for backwards compatibility, but removed from the RSDL metadata.

Change-Id: I9bb8dba2385b1219f6d90fa99692a79fc7cb36b7
Bug-Url: https://bugzilla.redhat.com/1136061
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>