core: Reduce # of calls to getVdsGroupByVdsGroupId
Since vds view is built on top of the vds_groups table (in addition to
other tables), this patch adds to vds the enable_ballon property so that
we don't need to load the full VdsGroup to get this value.

This saves ~90% of calls to getVdsGroupByVdsGroupId in an idle scenario,
each of these calls takes ~300us (depending on the number of clusters a
customer has).

Change-Id: I9d838a19e0496951b9821936bd3af215e3a31c0d
Signed-off-by: lzelkha@redhat.com <lzelkha@redhat.com>