core: setupNetworks not able to configure bridge
For 2 calls we use ThreadPoolUtil. It was used to provide async network
call for xmlrpc. Jsonrpc is async so from its prespective it makes no
sense to use it. In order to keep the communication interface unchanged
we need to returned Future.

The future checked whether the call was executed not whether the
response arrived. This patch makes sure that we wait for response and
expose this information correctly.


Bug-Url: https://bugzilla.redhat.com/1136876
Change-Id: I70bfa5d7f154a13f7d52f7710eb109de1da0b82c
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>