upload stream fails due to explicit cast
Code assumes that we all the time use xmlrpc and explicitly casts it to
VdsServerWrapper to get http client. When host uses jsonrpc it fails
with class cast excpetion.

This patch uses existing code to build http client and passes it to
JsonRpcVdsServer implementation. getHttpClient method is promoted to
IVdsServer interface so jsonrpc implementation can expose http client as
well.

Bug-Url: https://bugzilla.redhat.com/1138986
Change-Id: I81c1dbb50727172647da10ec4b74f2950057c4bb
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>