webadmin: The text and buttons not aligned in Create VM dialog
The http://gerrit.ovirt.org/#/c/30821 enhanced the "Guide Me" dialog so also
the bigger text will fit in. The problem with this patch is that it uses the
"display: flex" which is not supported by firefox 17 and firefox 17 is the only
browser supported by oVirt.

So this patch "reverts" this fix and applyes the one proposed here:
http://gerrit.ovirt.org/#/c/30861/ which does not look so nice but works also
on older browsers.

This patch also contains the http://gerrit.ovirt.org/#/c/32830/ which moves the
arrow in the login button of the iSCSI domain to the middle.

Change-Id: Ifcb74f036476fb8544d1f15d26a0ef92e2eeed46
Bug-Url: https://bugzilla.redhat.com/1135965
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>