core: allowed to add network label even if Vm using this network is running.
~ Network label was checked to be unchanged in method
'onlyPermittedFieldsChanged'. This is extracted out into new method
'allowedNetworkLabelManipulation' which returns true if network label
was not changed, or if former label was not specified (null) and new
one is specified (!=null).

~ moved methods out from non static SyncNetworkParametersBuilder.
These methods can be used outside from SyncNetworkParametersBuilder,
but in that case their owning class have to be instantiated using
CommandContext parameter, which is completely unrelated to these four
methods.

Change-Id: I40075be15ec9f3607c2f859b09e2b113d92f6e66
Bug-Url: https://bugzilla.redhat.com/1142203
Signed-off-by: Martin Mucha <mmucha@redhat.com>