core: fixing fencing flow for blocked proxy
This patch fixes the following scenario :

DC1 with H1 (with PM) and H2 on cluster C1
another host H3 on DC1 cluster C2

When we are blocking the communication with iptables from H2 to H1 PM
card and use the default proxy preferences (cluster, dc) a Restart
operation will always fail.

The fix relies on the return code from the script that was executed by
VDSM rather than the proceedProxyReturnValue call that throws an
exception and also set the execution result according to that code.

In addition a missing initialization of the 'count' variable in the
'for' loop has been added.

This BZ depends also in a VDSM fix that makes stop/start operation sync
instead of fire-and-forget.

Change-Id: I70d35967554c88c943c4653a26642e6e21cc54d2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1129381
Signed-off-by: Eli Mesika <emesika@redhat.com>