core: Change cluster command should lock host
This patch adds a missing lock when host cluster is changed.
Prior to this change the host was not locked allowing a concurrent
removal of the host to occur while the change cluster operation is still
running this resulted on a SQL exception while trying to add a line to
host SPM mapping table with a host GUID when the host was already
removed.

Change-Id: I1cd9e237838ce03438d806160d08efc9db0d6cd7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1102729
Signed-off-by: Eli Mesika <emesika@redhat.com>