core: Override unregistered entity on deatch.
Override unregistered entity when there is an existing entity which
already exists as an unregistered entity.

This can happen when a domain with an unregistered entity, is attached and
a VM with disks on multiple storage domains exists in the setup and also
as an unregistered VM.

Once the VM will be created with disk on the attached Storage Domain,
the unregistered entity should be overriden with the existing one, on
detach process.

Change-Id: Id22390ae392e95bbc73a5976fc886ed21da4989b
Bug-Url: https://bugzilla.redhat.com/1132568
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>