core: ImportVmTemplateCommand storage allocation
This patch is a part of a series of patches, adding storage allocation
validations to the system when they're missing, and replacing old
verification usage with unified, new, correct and tested verification.
This patch did this for ImportVmTemplateCommand, using only existing
validations. Removing old verification in this command resulted in
unused old validation and validation aids, which were also removed in
this patch.

Added tests to MoveOrCopyTemplateCommand.validateSpaceRequirements()
in MoveMoveOrCopyTemplateCommandTest.

Change-Id: I6ffc5c00905bb03c9cb3cbffa481e78c9b0ab87a
Bug-Url: https://bugzilla.redhat.com/1136721
Signed-off-by: Vered Volansky <vvolansk@redhat.com>