restapi: Introduce current filter
The RESTAPI application needs to access data associated to the current
request in several places. In order to simplify this we currently use a
"Current" class that handles a map stored in a thread local. This
"Current" class populates the map before processing each request, and
cleans it after finishing processing it. To do so it is defined as a
Resteasy interceptor. This interceptor concept isn't standard, and
doesn't work the same in newer versions of Resteasy. To be able to use
other versions of Resteasy this patch replaces the interceptor with a
plain servlet filter.

Change-Id: If46bafba6e8b58624af7eb85ae231edb0cf8dc1f
Bug-Url: https://bugzilla.redhat.com/1121327
Signed-off-by: Martin Perina <mperina@redhat.com>