engine,webadmin: Replace references to NetworkQoS with HostNetworkQos
Since now HostNetworkQos is a separate entity from NetworkQoS (as it
has different attributes), this updates the relevant references to the
entity itself and to its corresponding DAO and validator, as well as
related frontend dialogs.

Of course some logic had to be altered due to the difference between
NetworkQoS and HostNetworkQos, and similarly since dialog layouts have
changed since 3.4 they had to be updated.

Change-Id: Ia6a625d8c5b554cb525efc41d31013df0de2a146
Signed-off-by: Lior Vernia <lvernia@redhat.com>