core: AddVmTemplateCommand storage allocation
This patch is a part of a series of patches, adding storage allocation
validations to the system when they're missing, and replacing old
verification usage with unified, new, correct and tested verification.
This patch did this for AddVmTemplateCommand, using only existing
validations. Removing old verification in this command resulted in
unused old validation and validation aids, which were also removed in
this patch.

Change-Id: I7980510a7bb72e43e0a9dfc18460207386eb62fe
Bug-Url: https://bugzilla.redhat.com/1053744
Signed-off-by: Vered Volansky <vvolansk@redhat.com>