core, engine, userportal, webadmin: Block floppy support on PPC64
The ppc64 architecture does not support floppy devices, this change
blocks the use of such devices in VMs that do not support them.

A additional property in the osinfo (devices.floppy.support) was created
to indicate if the guest OS supports floppy devices. The RunVmCommand
was changed to validate if the selected VM can use the attached floppy
image and the frontend was changed to hide floppy-related options if
the VM does not support it.

Change-Id: I8fb783bbafa3c6a18e327e4e02aac3d71b6f110b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1067605
Signed-off-by: Vitor de Lima <vdelima@redhat.com>