core: Validate Storage Domain for Importing Template from configuration
For each disk in the template, we should fetch the Storage Domain and
validate that it exists and active.
Today, we set the storage id for each image and then use
getStorageDomain method to fetch
the Storage Domain with that id, but when the Storage Domain being
initialized, it will always check
the first Storage Domain every time.

Change-Id: I2bfe3ba9b48747650291a887255e85358ed0799a
Bug-Url: https://bugzilla.redhat.com/1138200
Bug-Url: https://bugzilla.redhat.com/1140560
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>