webadmin: Refactor QoS model side
The hierarchy was overly-complex; the new and edit branches of the
hierarchy tree were hardly any different and did not merit
separation. The logic of the separation was also inconsistent, as
executeSave() was shared but performed operations the edit branch
didn't require, while the init() was mutually exclusive.

Also added validation for QosParametersModel as part of the generic
QosModel flow.

Change-Id: I2374a54fbe1dc744b71071869b4a23ebc100acea
Signed-off-by: Lior Vernia <lvernia@redhat.com>