core: Don't write empty description/comment to OVF
Currently when we generate OVF files we write the description and
comment properties of the VM even if they are null. As a result when a
VM without a description or comment is re-imported it has an empty
string a value (instead of null) for these properties. This patch
changes the OVF writer so that it won't write the description or comment
properties if they are null.

Change-Id: Iffd0f7010bc32402c45882dc6087bf02a55a14d9
Bug-Url: https://bugzilla.redhat.com/1145466
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>