core: Remove Memory and OVF Volumes on VM Remove
When removing a VM that has a snapshot with memory, its memory and OVF
volumes are not removed. The reason for this is the value of the
property removeOnlyIfNotUsedAtAll in RemoveMemoryVolumesParameters.
If we don't set its value to true, by default it's false and then we can
remove the memory and OVF volumes before the VM's disks.

Change-Id: I5303c571e3d8843c03b195ec91763279053bc5de
Bug-Url: https://bugzilla.redhat.com/1145073
Bug-Url: https://bugzilla.redhat.com/1145188
Bug-Url: https://bugzilla.redhat.com/1147085
Bug-Url: https://bugzilla.redhat.com/1145254
Signed-off-by: Idan Shaby <ishaby@redhat.com>