core: medium Instance Type mismatch feature page
According to feature page, in medium instance type, total virtual CPUs should
be 2, but it is 1.

Fixed by updating this property in the DB. The update statement tries to be as
specific about the instance type as possible, but since it is not possible to
use it's ID which was generated, it relyes on the name/description + some
configuration.

Change-Id: I03407f86c272578cec671f49e89d2ffbeb93ae4d
Bug-Url: https://bugzilla.redhat.com/1142924
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>