core: retry fencing operation in case of failure
retry fencing operation in case of a temporary failure

Up to now, if 'unknown' was returned while waiting for a certain status
('on' / 'off') , the while loop was exited with a proper message,
breaking the reboot sequence.
It seems that there are some cases in which a fencing operation may fail
temporarily and succeeded after a while.
This patch retries to get to the desired status 3 times before giving up
with the configured delay between each status check.

Change-Id: Iab5dadd5cb07f2c61dfdf6d75e9fccca790ba483
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1140098
Signed-off-by: Eli Mesika <emesika@redhat.com>