core: AddVmPoolWithVmsCommand storage allocation
This patch is a part of a series of patches, adding storage allocation
validations to the system when they're missing, and replacing old
verification usage with unified, new, correct and tested verification.
This patch did this for AddVmPoolWithVmsCommand, using only existing
validations.
Previously the storage allocation check was wrongfully in
CommonVmPoolWithVmsCommand, and is now extracted to
AddVmPoolWithVmsCommand only.

Change-Id: I3120acf7eb5c71be6ade437522174a7bf62ea52a
Bug-Url: https://bugzilla.redhat.com/1143888
Signed-off-by: Vered Volansky <vvolansk@redhat.com>