core: Merge snapshots storage allocation
This patch is a part of a series of patches, adding storage allocation
validations to the system when they're missing, and replacing old
verification usage with unified, new, correct and tested verification.
This patch did this for RemoveSnapshotCommand and
RemoveDiskSnapshotsCommand.
This also resulted in removing tests of old/buggy verification.

Change-Id: I8081eaf60af50ad8d894454244fb709f4a08d5c6
Bug-Url: https://bugzilla.redhat.com/1053733
Signed-off-by: Vered Volansky <vvolansk@redhat.com>