core: allow to maintenance host if install failed
Allow to put host in maintenance if install failed

This patch allows to put a host in maintenance in case that host
installation failed.
In networking problems it may occur that after trying to put the host in
maintenance it will go to preparing for maintenance=>Non Responding and
we will have to put it in maintenance again.

Change-Id: Ic3611e4fb7ac039ee4014f6cb158b0da5067331e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1147988
Signed-off-by: Eli Mesika <emesika@redhat.com>