core: fix attaching a cluster to a new DC
This patch fixes attaching a cluster to a new DC after its original DC
was removed.

In that case if the cluster has a host it will loss its dc id settings
and will try to insert to the vds_spm_map table empty GUID as the DC id.
Code was changed to check if the host DC id is empty and if so to take
this value from the cluster.

Change-Id: I195b333d8d8fafd0b21831961862d6dc2ca42ee9
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1148301
Signed-off-by: Eli Mesika <emesika@redhat.com>