restapi: Validate QoS type
Currently when a QoS is added to a data center we don't validate that
the mandatory type is specified. This eventually triggers a NPE when
validating the value. To avoid this this patch adds the validation and
modifies the type validator so that it won't fail if no type is
provided.

Change-Id: I90b62c4f4499dcc4abec5495a9907102e213df0e
Bug-Url: https://bugzilla.redhat.com/1148773
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit ba3dbd7a4e7e8f516445a9717c9f7a74ddbfe90b)