restapi: Fix use of "unregistered" parameter
The "unregister" matrix parameter is used to include unregistered
objects (disks, templates and VMs) when listing or adding them. But the
paremeter is currently documented incorrectly, as it is missing from
some operations where it is used and included in other operations where
it isn't used. As a result the RSDL is incorrect and the Python and Java
SDKs don't implement it correctly. This patch fixes the documentation of
the parameter.

Change-Id: Ifccc9082d71cc6bb14ae9dfc7efd3c024a82639f
Bug-Url: https://bugzilla.redhat.com/1148715
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 29012b25c07d964d8f7aeeba7b6703bf452ae6b9)