core: RunVmValidator storage allocation
This patch is a part of a series of patches, adding storage allocation
validations to the system when they're missing, and replacing old
verification usage with unified, new, correct and tested verification.
This patch did this for RunVmValidator.
In doing so, space allocation validation for stateless VM became
redundant and was therefore removed.

Change-Id: Ib3f7c7f31d738f84d356d4e89b05a21d4ca7a118
Bug-Url: https://bugzilla.redhat.com/1149809
Signed-off-by: Vered Volansky <vvolansk@redhat.com>