core: Import image from glance storage allocation
This patch is a part of a series of patches, adding storage allocation
validations to the system when they're missing, and replacing old
verification usage with unified, new, correct and tested verification.
This patch did this for ImportRepoImageCommand, using only existing
validations.
Reorganized sparse/preallocated determination as part of the CDA flow as
to have the correct storage allocation validation.

Change-Id: I60b7f13d8f64608f97e4eeebe2f9eef3087d9920
Bug-Url: https://bugzilla.redhat.com/1053750
Signed-off-by: Vered Volansky <vvolansk@redhat.com>