webadmin: Added Qos per profile when removing qos
When removing multiple qos for datacenter which are connected to
storage profiles a message appears which state which profiles are
connected to those qos, but without stating which qos belong
to each profile(This is a problem only when selecting multiple qos
for removal).

The following fix adds the associated qos to each
profile name in the message.

Change-Id: I474d38f934f8aaef482f06a81603bbd8566cb871
Bug-Url: https://bugzilla.redhat.com/1142854
Signed-off-by: Tomer Saban <tsaban@redhat.com>