core, frontend: Query to get raw provider certificates
Currently we have a query that retrieves the certificates of an external
provider converted to a string, and this string is displayed to the
user. The RESTAPI can't easily consume this text, it should rather get
the raw certificates and render them so that callers can decide what
information to extract. To make this possible this patch renames the
existing query, to make it clear that it returns just the text, and adds
a new one that returns the raw certificates, intended for use by the
RESTAPI.

Change-Id: Ie5270bff9b671943fdea1c31369b7ad1d0dd4f67
Related-To: https://bugzilla.redhat.com/1132259
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>