restapi: Fix NPE in request for CDROMs of wrong VM
Currently when we receive a request for the collection of CDROMs of a VM
we don't take into account that the query performed to find the VM can
return null if there isn't such a VM (the identifier is wrong, or the VM
has been removed). To avoid that NPE this patch checks the result of the
query and will respond with HTTP code 404 (not found) in those cases.

Change-Id: I55d4742649d8d260ae4d7c924c1a9b91efbcea6b
Bug-Url: https://bugzilla.redhat.com/1150953
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 8d5aa3646e2402a48f88a0532a70d9bb50347e57)