core: Update commons-collections to 3.2.1
Currently we are using version 3.1 of commons-collections in our root
POM, but we are actually using 3.2.1 as that is what is provided by the
application server. This means that we can't use some of the methods
introduced in version 3.2, like "isEmpty", but only because the build
will fail. This patch updates the version number in the root POM to
3.2.1 so that we can use those methods.

Change-Id: I68ed14360d61aa3d106fff2b690f7dda0507303f
Related-To: https://bugzilla.redhat.com/1150953
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>