core: dont copy old num of sockets on hotplug
when editing running vm, all fields that are not editable, are copied
from the old vm to new vm, as they are saved in the running snapshot,
and db reflect the actual running config.

when user select to do hot-set to cpu, the number of sockets should not
be copied with the old value, as it is used in the hot-set command.

this is fixed by sending to the copy method, if this is a hot-set flow,
and it will check it when doing copy of the old value, for hot-set allowed fields.

Change-Id: I57bcd2287350c3146f06fe2a5c9a583c88dded04
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1151827
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>