Allow to avoid lock screen on spice disconnect
This patch introduces a feature to be able to select the type of action
to perform, when an user disconnects from the spice console.

Previously this was always trying to perform 'lock screen' via the guest
agent. Now this can be disabled, the user can be logged out, the VM
rebooted or shutdown.

Change-Id: Ia2ef5ffaceed619f6630b56a7156f25e9111fd9e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1171638
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>