engine: Enforce maximal QoS commitment on interface
Made sure that no more than 75% of a an interface's capacity can be
committed by QoS, as overcommitment may lead to unexpected and
potentially destructive behavior.

As part of this validation, also made sure all interface speeds are
either reported by vdsm or can be computed in the engine
(newly-created bonds); otherwise fail, as we cannot guarantee that the
user hasn't over-committed.

Change-Id: Icc89f4193d6b7b590c6e1e053af74ab23bc77a11
Signed-off-by: Lior Vernia <lvernia@redhat.com>