core: RemoveDiskSnapshots - sort images from parent to leaf
When removing multiple disk snapshots that belongs to the same disk,
the images should be sorted from parent to leaf (active).
Sorting is needed since removing a child image and a parent image
afterwards, might be invalid - as the parent will point to another child
and when deletion is done in consecutive order there's a race condition
between DB update and removing next parent.

Change-Id: I759537f7c32fc9b19a968f32767bba633f3f3366
Bug-Url: https://bugzilla.redhat.com/1151841
Signed-off-by: Daniel Erez <derez@redhat.com>