core: use previous disk operator on template version update
Updating the template version for a VM when the template contained disks
didn't work because no user was selected to serve as the disks operator.

This patch fixes it by setting the user that used to be disk operator
for one of the disks of the VM while it was based on the previous
template version, as the disk operator of the disks from the new
template version.

Change-Id: Idd9604c7a219e49d140576a66c9b15e9abb34d79
Bug-Url: https://bugzilla.redhat.com/1149135
Signed-off-by: Arik Hadas <ahadas@redhat.com>