core: Use Jackson 1.9.9
The version of the OpenStack Java SDK that we use requires at least
Jackson 1.9.9 in order to work correctly, otherwise responses received
from Neutron aren't parsed correctly. This patch changes the
dependencies module so that the correct version of Jackson will be
bundled with the engine, overriding the one provided by the
application server.

Change-Id: I55d02ecd385c86a32a69f154610d4093aea9b74e
Bug-Url: https://bugzilla.redhat.com/1064231
Related-To: https://bugzilla.redhat.com/1132259
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>