core: use MDC to hold correlation id
apparently the slf4j within jboss and MDC support of the log handler
are integrated, so we can levarage that in order to delegate correlation
id to log without actually have our own logger.

all methods of getting correlation id out of this context seems to be
presentation only usage which is correct.

Bug-Url: https://bugzilla.redhat.com/1109871
Change-Id: Iff53d3bf420e6fe1b1a0a04a0a21a49a3c01a4d2
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>