core: copy vm-init data on new template version
when creating new template version, the vm-init data is not copied,
because the vmInit is not marked for copy.

since the vm-init is saved in a different table, it need to be
manually loaded.

also, the is_initialized flag need to be reset so vms will do sysprep
again with new data.

Change-Id: Icaf1b6cad241008c8e6f392cb918e900b340d9b2
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1148379
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>