core: support custom fencing params for PPC
Adding support for PPC hosts in 3.5 raised the need of having separate
default fencing parameters for PPC.

Default parameters are actually parameters with certain values per agent
that are sent implicitly to VDSM when the agent is used.

The problem is that those parameters that insures the correctness of the
call to the agent script may vary between X86 and PPC

This patch introduces a separate fencing default parameters for PPC and
an option to customize it for a custom agent as well.

Please note that the architecture checked is of the target fenced host for the
operation.

Change-Id: I89717ac346e2ca1b44021b18dbe989ca5399f3f5
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1149235
Signed-off-by: Eli Mesika <emesika@redhat.com>