core: Fix mapping of US Eastern Standard Time Zone
The WindowsJavaTimezoneMapping mapped US East. ST to java time zone
GMT+5 which was not adjusted to Daylight Savings Time.
Now we map it to America/New_York which takes DST into account.

Also fixed another occurence of mapping to raw GMT timezone:
SA Eastern Standard Time to America/Buenos_Aires (was GMT+3) but this
change was not as consequential since Argentina does not have DST.
Although this should be more resilient if they decided to use it in the
future.

Similar story for Georia -> Asia/Tbilisi.

Change-Id: I0d17376caac96d1aae87cc894452d42e161637e4
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1148623
Signed-off-by: Martin Betak <mbetak@redhat.com>