core: Store iSCSI Target Address instead of IP Address in DB
When discovering a new iSCSI target, we send VDSM the address
we got from the user's input (which can be an IP address or a DNS name).
Since VDSM translates any address to an IP address and sends it back to
the engine, we don't want to store this address in the DB, but the original
one that we got from the user.

Change-Id: Idf6d385606e130f4f529a0b374a9f731310a7f9f
Bug-Url: https://bugzilla.redhat.com/1086310
Signed-off-by: Idan Shaby <ishaby@redhat.com>