webadmin: Maximum number of VMs per user in VM Pool dialog validated
The problem was that validation of 'Maximum number of VMs per user'
field in both 'New pool' and 'Edit pool' dialog was simple interval
[1, 32767]. So it was possible to set Max num of VMs per user higher
than total number of available VMs in pool.

Now the upper bound of Max num of VMs per user is total number of VMs
in pool. Fixed in PoolModelBehaviorBase#validate(),
CommonApplicationMessages#maxAssignedVmsPerUserHelp()

Change-Id: Ic70d45b5e8d84965691777430f8f362a68bd4830
Bug-Url: https://bugzilla.redhat.com/1124360
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>