webadmin: CPU shares custom value is now possible
On the "New Virtual Machine" dialog, inside the "Resource Allocation"
tab on the left, when selecting "Custom" on the CPU shares combobox,
the textbox on the right(Where you set the actual value of the CPU
shares) wasn't changeable.
Now, it becomes changeable once "Custom" is chosen and unchangeable
once another value is chosen in the combobox.

Change-Id: Ie1fe20afa682926ac14b7d9ac7c5fa589b3b0459
Bug-Url: https://bugzilla.redhat.com/1152636
Signed-off-by: Tomer Saban <tsaban@redhat.com>