core, engine, restapi: Proper default migration policies for ppc64
This patch changes the default option when creating clusters and VMs
using the REST API. When omitted, the Migrate On Error parameter for
new clusters is disabled when the architecture of the cluster does not
support migration. Also, if the user does not specify a migration
policy for a new VM, the REST API will disable it in architectures that
do not support it.

Change-Id: Ib34c4a01fe0c667bafc47bf70b33bb2990ffb7d3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1057178
Signed-off-by: Vitor de Lima <vdelima@redhat.com>